+++ KeyFinderLib/KeySearchTypes.h
@@ -32,7 +32,7 @@ typedef struct {
     std::string deviceName;
     uint64_t freeMemory;
     uint64_t deviceMemory;
-    size_t targets;
+    uint64_t targets;
     secp256k1::uint256 nextKey;
 }KeySearchStatus;
